+++ medicare_dbt/models/marts/mart_enrollment_national.sql
@@ -30,8 +30,8 @@ national as (
         disabled_beneficiaries,
         male_beneficiaries,
         female_beneficiaries,
-        ROUND(medicare_advantage_benes / NULLIF(total_beneficiaries, 0) * 100, 1) AS ma_penetration_rate,
-        ROUND(original_medicare_benes / NULLIF(total_beneficiaries, 0) * 100, 1)  AS ffs_penetration_rate,
+        ROUND((medicare_advantage_benes / NULLIF(total_beneficiaries, 0) * 100)::numeric, 1) AS ma_penetration_rate,
+        ROUND((original_medicare_benes / NULLIF(total_beneficiaries, 0) * 100)::numeric, 1)  AS ffs_penetration_rate,
         year,
         month
       from staging
